- XUnits   - Define financial values as Decimal.   - Run all tests and document. - Maui App   - Update.

+++ src/sep.ira/calculator.xunits/AppreciatorCalculationTests.cs
@@ -62,9 +62,9 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueInitialTaxIsAppliedCorrectly()
     {
         // Arrange
-        double investedAmount = 100000;
-        double initialFederalTax = 20;
-        double initialStateTax = 10;
+        decimal investedAmount = 100000;
+        decimal initialFederalTax = 20;
+        decimal initialStateTax = 10;
         this._appreciator.InvestedAmount = investedAmount;
         this._appreciator.InitialFederalTaxRate = initialFederalTax;
         this._appreciator.InitialStateTaxRate = initialStateTax;
@@ -75,7 +75,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        double expectedInitialCapital = investedAmount * (1 - ((initialFederalTax + initialStateTax) / 100.0));
+        decimal expectedInitialCapital = investedAmount * (1 - ((initialFederalTax + initialStateTax) / 100.0m));
         Assert.Equal( expectedInitialCapital, this._appreciator.CapitalAccountBalance, 2 );
     }
 
@@ -87,7 +87,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithHighGrowthRateProducesLargeGain()
     {
         // Arrange
-        double investedAmount = 50000;
+        decimal investedAmount = 50000;
         this._appreciator.InvestedAmount = investedAmount;
         this._appreciator.AnnualGrowthRate = 15;  // High growth rate
         this._appreciator.InitialFederalTaxRate = 0;
@@ -98,7 +98,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        double expectedBalance = investedAmount * Math.Pow( 1.15, 20 );
+        decimal expectedBalance = investedAmount * ( decimal ) Math.Pow( 1.15, 20 );
         Assert.Equal( expectedBalance, this._appreciator.CapitalAccountBalance, 0 );
         Assert.True( this._appreciator.CapitalGain > investedAmount * 2,
             "With 15% growth over 20 years, capital should more than triple" );
@@ -138,7 +138,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithMultipleYearsComputesCorrectly()
     {
         // Arrange
-        double investedAmount = 10000;
+        decimal investedAmount = 10000;
         this._appreciator.InvestedAmount = investedAmount;
         this._appreciator.AnnualGrowthRate = 7;
         this._appreciator.InitialFederalTaxRate = 0;
@@ -149,8 +149,8 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        double growthRate = 1.07;
-        double expectedBalance = investedAmount * Math.Pow( growthRate, 5 );
+        decimal growthRate = 1.07m;
+        decimal expectedBalance = investedAmount * ( decimal ) Math.Pow( ( double ) growthRate, 5 );
         Assert.Equal( expectedBalance, this._appreciator.CapitalAccountBalance, 2 );
     }
 
@@ -162,7 +162,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithNegativeGrowthReducesCapital()
     {
         // Arrange
-        double investedAmount = 100000;
+        decimal investedAmount = 100000;
         this._appreciator.InvestedAmount = investedAmount;
         this._appreciator.AnnualGrowthRate = -5;  // Negative growth
         this._appreciator.InitialFederalTaxRate = 0;
@@ -188,7 +188,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithZeroInvestmentDurationKeepsInitialAmount()
     {
         // Arrange
-        double investedAmount = 50000;
+        decimal investedAmount = 50000;
         this._appreciator.InvestedAmount = investedAmount;
         this._appreciator.InvestmentDuration = 0;
         this._appreciator.InitialFederalTaxRate = 0;
@@ -274,8 +274,8 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValueSepIraWithRmd();
 
         // Assert - RMD at age 72 from table is 27.4
-        double expectedRmd = this._appreciator.InvestedAmount / 27.4;
-        double expectedSepiIraBalance = this._appreciator.InvestedAmount - expectedRmd;
+        decimal expectedRmd = this._appreciator.InvestedAmount / 27.4m;
+        decimal expectedSepiIraBalance = this._appreciator.InvestedAmount - expectedRmd;
         Assert.True( this._appreciator.CapitalAccountBalance > 0,
             "Capital account should have RMD contribution" );
         Assert.True( this._appreciator.SepIraAccountBalance < this._appreciator.InvestedAmount,
@@ -359,7 +359,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValueSepIraWithRmd();
 
         // Assert
-        double sepPlusCapital = this._appreciator.SepIraAccountBalance + this._appreciator.CapitalAccountBalance;
+        decimal sepPlusCapital = this._appreciator.SepIraAccountBalance + this._appreciator.CapitalAccountBalance;
         Assert.True( sepPlusCapital > this._appreciator.InvestedAmount,
             "With growth, total should exceed initial investment" );
     }
@@ -397,8 +397,8 @@ public class AppreciatorCalculationTests
         // Assert - check that the table has expected ranges
         Assert.Contains( 72, this._appreciator.UniformLifetimeTable.Keys );
         Assert.Contains( 120, this._appreciator.UniformLifetimeTable.Keys );
-        Assert.Equal( 27.4, this._appreciator.UniformLifetimeTable[72], 1 );
-        Assert.Equal( 1.9, this._appreciator.UniformLifetimeTable[120], 1 );
+        Assert.Equal( 27.4m, this._appreciator.UniformLifetimeTable[72], 1 );
+        Assert.Equal( 1.9m, this._appreciator.UniformLifetimeTable[120], 1 );
     }
 
     /// <summary>   (Unit Test Method) uniform lifetime table values decrease with age. </summary>
@@ -410,8 +410,8 @@ public class AppreciatorCalculationTests
         // Note: IRS table has a special case where 105 and 106 both have 4.5
         for ( int age = 72; age < 120; age++ )
         {
-            if ( this._appreciator.UniformLifetimeTable.TryGetValue( age, out double atAge ) &&
-                 this._appreciator.UniformLifetimeTable.TryGetValue( age + 1, out double atAgePlus1 ) )
+            if ( this._appreciator.UniformLifetimeTable.TryGetValue( age, out decimal atAge ) &&
+                 this._appreciator.UniformLifetimeTable.TryGetValue( age + 1, out decimal atAgePlus1 ) )
             {
                 // Most ages: divisor decreases
                 // Special case at 105->106: divisor stays the same (both 4.5)
@@ -433,7 +433,7 @@ public class AppreciatorCalculationTests
     public void InvestedAmountPropertyCanBeSetAndRetrieved()
     {
         // Arrange
-        double testAmount = 250000;
+        decimal testAmount = 250000;
 
         // Act
         this._appreciator.InvestedAmount = testAmount;
@@ -502,12 +502,12 @@ public class AppreciatorCalculationTests
     public void EconomicRatePropertiesCanBeSetAndRetrieved()
     {
         // Arrange & Act
-        this._appreciator.AnnualInflationRate = 3.5;
-        this._appreciator.AnnualGrowthRate = 8.5;
+        this._appreciator.AnnualInflationRate = 3.5m;
+        this._appreciator.AnnualGrowthRate = 8.5m;
 
         // Assert
-        Assert.Equal( 3.5, this._appreciator.AnnualInflationRate );
-        Assert.Equal( 8.5, this._appreciator.AnnualGrowthRate );
+        Assert.Equal( 3.5m, this._appreciator.AnnualInflationRate );
+        Assert.Equal( 8.5m, this._appreciator.AnnualGrowthRate );
     }
 
     #endregion
@@ -523,7 +523,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithAllZeroTaxRatesAndZeroGrowthPreservesCapital()
     {
         // Arrange
-        double amount = 75000;
+        decimal amount = 75000;
         this._appreciator.InvestedAmount = amount;
         this._appreciator.InitialFederalTaxRate = 0;
         this._appreciator.InitialStateTaxRate = 0;
@@ -549,7 +549,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithVeryHighTaxRatesReducesCapitalSignificantly()
     {
         // Arrange
-        double amount = 100000;
+        decimal amount = 100000;
         this._appreciator.InvestedAmount = amount;
         this._appreciator.InitialFederalTaxRate = 50;
         this._appreciator.InitialStateTaxRate = 50;
@@ -571,7 +571,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithSmallInvestmentProducesProportionalResults()
     {
         // Arrange
-        double smallAmount = 100;
+        decimal smallAmount = 100;
         this._appreciator.InvestedAmount = smallAmount;
         this._appreciator.AnnualGrowthRate = 10;
         this._appreciator.InitialFederalTaxRate = 0;
@@ -582,7 +582,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        Assert.Equal( smallAmount * 1.10, this._appreciator.CapitalAccountBalance, 2 );
+        Assert.Equal( smallAmount * 1.10m, this._appreciator.CapitalAccountBalance, 2 );
     }
 
     /// <summary>
@@ -594,7 +594,7 @@ public class AppreciatorCalculationTests
     public void CalculateFutureValueWithLargeInvestmentProducesProportionalResults()
     {
         // Arrange
-        double largeAmount = 10000000;
+        decimal largeAmount = 10000000;
         this._appreciator.InvestedAmount = largeAmount;
         this._appreciator.AnnualGrowthRate = 5;
         this._appreciator.InitialFederalTaxRate = 0;
@@ -605,7 +605,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        Assert.Equal( largeAmount * 1.05, this._appreciator.CapitalAccountBalance, 0 );
+        Assert.Equal( largeAmount * 1.05m, this._appreciator.CapitalAccountBalance, 0 );
     }
 
     #endregion
@@ -650,9 +650,9 @@ public class AppreciatorCalculationTests
         sepIraAppreciator.CalculateFutureValueSepIraWithRmd();
 
         // Assert - SEP IRA should produce at least as much (no RMDs until age 72)
-        double simpleFutureValue = simpleAppreciator.CapitalAccountBalance;
-        double sepIraFutureValue = sepIraAppreciator.SepIraAccountBalance + sepIraAppreciator.CapitalAccountBalance;
-        Assert.True( sepIraFutureValue >= simpleFutureValue * 0.95,
+        decimal simpleFutureValue = simpleAppreciator.CapitalAccountBalance;
+        decimal sepIraFutureValue = sepIraAppreciator.SepIraAccountBalance + sepIraAppreciator.CapitalAccountBalance;
+        Assert.True( sepIraFutureValue >= simpleFutureValue * 0.95m,
             "SEP IRA should produce comparable results" );
     }
 
@@ -669,9 +669,9 @@ public class AppreciatorCalculationTests
         Appreciator newAppreciator = new();
 
         // Assert - output properties should be initialized (may be 0 or default)
-        _ = Assert.IsType<double>( newAppreciator.SepIraAccountBalance );
-        _ = Assert.IsType<double>( newAppreciator.CapitalAccountBalance );
-        _ = Assert.IsType<double>( newAppreciator.WithdrawalTaxLiability );
+        _ = Assert.IsType<decimal>( newAppreciator.SepIraAccountBalance );
+        _ = Assert.IsType<decimal>( newAppreciator.CapitalAccountBalance );
+        _ = Assert.IsType<decimal>( newAppreciator.WithdrawalTaxLiability );
     }
 
     /// <summary>
@@ -693,7 +693,7 @@ public class AppreciatorCalculationTests
         this._appreciator.CalculateFutureValue();
 
         // Assert
-        double expectedNetValue = this._appreciator.CapitalAccountBalance - this._appreciator.WithdrawalTaxLiability;
+        decimal expectedNetValue = this._appreciator.CapitalAccountBalance - this._appreciator.WithdrawalTaxLiability;
         Assert.Equal( expectedNetValue, this._appreciator.NetCashOutValue, 2 );
     }
